Fix rubocop warning.

+++ providers/default.rb
@@ -253,9 +253,7 @@ end
 def prepare_windows
   copy_opt_chef(chef_install_dir, chef_backup_dir)
   Kernel.spawn("c:/windows/system32/schtasks.exe /F /RU SYSTEM /create /sc once /ST \"#{upgrade_start_time}\" /tn Chef_upgrade /tr \"powershell.exe -ExecutionPolicy Bypass c:/opscode/chef_upgrade.ps1\"")
-  if allow_exit?
-    FileUtils.rm_rf "#{chef_install_dir}/bin/chef-client.bat"
-  end
+  FileUtils.rm_rf "#{chef_install_dir}/bin/chef-client.bat" if allow_exit?
 end
 
 def uninstall_ps_code
